Created several "Multi" objects in two different attempts for this
I'm not sure if it's possible to pass multiple users RESTfully, as the Simulator object uses the Web/ folder for all it's work. This may still be circumvented by use of the State object. The other option I am trying to cobble together is to just run several servers with different ports, and a handler to 403 redirect, but at present niether method is functional.